Fix ref on stateless function component warning
The `withTheme` HOC spec was passing a stateless function component,
which was raising this warning:

"Warning: Stateless function components cannot be given refs.
 Attempts to access this ref will fail"

`withTheme` always needs to be passed a class-based component. It's
actually something I hadn't thought about, but we're fine in all live
components as they're either already a class, or wrapped with `Radium`.