Add Pill component (#212)
This PR adds the `Pill` Component to snacks. The one question I had was around the color API. Currently it will throw a proptype warning if it is not a color in our styleguide which would ensure we are staying on brand while using this component. This may not be what we want moving forwards @mmarcuccio @dcocchia thoughts?

<img width="1412" alt="screen shot 2018-08-21 at 11 06 18 pm" src="https://user-images.githubusercontent.com/3149916/44440613-e5919180-a596-11e8-94da-b3187bc576aa.png">