Fix incompatibility with react-hot-loader
Snacks components which depend on React element type comparison do not
function correctly when used from an environment running react-hot-loader.
This is a known issue with react-hot-loader and not necessarily an issue
in Snacks. However, since react-hot-loader is used so widely, we will
update snacks for now to patch this incompatibility.

* Add utility function for runtime element type comparisons
* Refactor Menu component to use new function

Fixes #235